+++ CHANGELOG.md
@@ -6,9 +6,13 @@
 
 ### Fixed
 
+ - Fixed assignment of TbProfiler variant type
+
 ### Changed
 
+ - Added reference, note fields to PhenotypeInfo
  - Mykrobe output parser handles csv format instead of json
+ - Split generic ResistanceVariant to one specified for individual tools
 
 ## [0.2.0]